+++ src/middlewares/auth.middleware.js
@@ -1,8 +1,30 @@
+// auth.middleware.js
+import jwt from "jsonwebtoken";
+import { User } from "../persistence/mongo/models/user.model.js";
+
 const JWT_SECRET = process.env.JWT_SECRET || "your_jwt_secret_key";
 
-export const isAuthenticated = (req, res, next) => {
-    if (req.isAuthenticated()) {
+export const jwtViewAuth = async (req, res, next) => {
+    const token = req.cookies?.token;
+
+    if (!token) {
+        res.locals.user = null;
         return next();
     }
-    return res.status(401).json({ message: "Not authorized, please log in" });
+
+    try {
+        const decoded = jwt.verify(token, JWT_SECRET);
+        const user = await User.findById(decoded.id).lean();
+
+        if (user) {
+            req.user = user;
+            res.locals.user = user; // makes it accessible to views
+        } else {
+            res.locals.user = null;
+        }
+    } catch (err) {
+        res.locals.user = null;
+    }
+
+    next();
 };
